Build code: Fix version. Move numbers outside of shortcode
Before this change users couldn't see clearly, when editing the page,
where are all the versions (online, core, branch). Some were in the
shortcode some were in the build-code.md. Best to have it all in one
place and visible from the get go (build-code.md)

Additionally: remove legacy /.autogen.sh and ./configure commands and
replace it with the existing build-online section

Signed-off-by: Pedro Pinto Silva <pedro.silva@collabora.com>